cleanup type casting, add noexcepts where needed
also moved some enums that will likely never need to be updated to their own headers, such as blips and horn names
