+++ Solution/source/Misc/MagnetGun.cpp
@@ -95,7 +95,7 @@ namespace MagnetGun
 
 	void MagnetGun::DisplayMagnetGunHelp()
 	{
-		Game::Print::PrintBottomLeft(oss_ << "Use the ~b~" << get_weapon_label(_whash, true), "~s~ to use the magnet ability.");
+		Game::Print::PrintBottomLeft(oss_ << "Use the ~b~" << GetWeaponLabel(_whash, true), "~s~ to use the magnet ability.");
 		Game::Print::PrintBottomLeft(oss_ << "Use the ~b~ " << (Menu::bitController ? "L and R Sticks" : "Mouse Scroll") << "~s~ to alter the magnet gun distance.");
 	}
