+++ Solution/source/Scripting/GTAplayer.h
@@ -24,11 +24,6 @@ class GTAplayer
 {
 private:
 	INT8 index;
-	/*bool exists, alive, isinvehicle;
-	ScrHandle ped, vehicle;
-	PCHAR name;
-	Hash weaponhash;
-	Vector3_t coords;*/
 
 public:
 	GTAplayer();
@@ -36,14 +31,12 @@ public:
 	GTAplayer(INT8 handle);
 
 	struct NetHandle13{ int _0, _1, _2, _3, _4, _5, _6, _7, _8, _9, _10, _11, _12; };
-	//struct ClanDesc35{ int _0, _1, _2, _3, _4, _5, _6, _7, _8, _9, _10, _11, _12, _13, _14, _15, _16, _17, _18, _19, _20, _21, _22, _23, _24, _25, _26, _27, _28, _29, _30, _31, _32, _33, _34; };
 
 	void SetIndex(INT8); /// set new pIndex for player
 
 	INT8& Handle(); /// return index reference
-	INT8 GetHandle() const; /// return index
+	INT8 GetHandle() const noexcept; /// return index
 	GTAped GetPed() const; /// return player's ped handle
-	//GTAvehicle GetVehicle() const; /// return player's vehicle handle
 	std::string GetName() const; /// return player's name
 	bool IsActive() const; /// return if exists
 	bool IsAlive() const; /// return true if alive
