+++ Solution/source/Scripting/GTAvehicleData.h
@@ -1,0 +1,128 @@
+/*
+* Menyoo PC - Grand Theft Auto V single-player trainer mod
+* Copyright (C) 2019  MAFINS
+*
+* This program is free software: you can redistribute it and/or modify
+* it under the terms of the GNU General Public License as published by
+* the Free Software Foundation, either version 3 of the License, or
+* (at your option) any later version.
+*/
+#pragma once
+
+#include <string>
+#include <vector>
+
+inline const std::vector<std::string> vValues_HornNames
+{
+	"CMOD_HRN_TRK",
+	"CMOD_HRN_COP",
+	"CMOD_HRN_CLO",
+	"CMOD_HRN_MUS1",
+	"CMOD_HRN_MUS2",
+	"CMOD_HRN_MUS3",
+	"CMOD_HRN_MUS4",
+	"CMOD_HRN_MUS5",
+	"CMOD_HRN_SAD",
+	"HORN_CLAS1",
+	"HORN_CLAS2",
+	"HORN_CLAS3",
+	"HORN_CLAS4",
+	"HORN_CLAS5",
+	"HORN_CLAS6",
+	"HORN_CLAS7",
+	"HORN_CNOTE_C0",
+	"HORN_CNOTE_D0",
+	"HORN_CNOTE_E0",
+	"HORN_CNOTE_F0",
+	"HORN_CNOTE_G0",
+	"HORN_CNOTE_A0",
+	"HORN_CNOTE_B0",
+	"HORN_CNOTE_C1",
+	"HORN_HIPS1",
+	"HORN_HIPS2",
+	"HORN_HIPS3",
+	"HORN_HIPS4",
+	"HORN_INDI_1",
+	"HORN_INDI_2",
+	"HORN_INDI_3",
+	"HORN_INDI_4",
+	"HORN_LUXE1",
+	"HORN_LUXE2",
+	"HORN_LUXE3",
+	"",
+	"",
+	"",
+	"HORN_HWEEN1",
+	"",
+	"HORN_HWEEN2",
+	"",
+	"HORN_LOWRDER1",
+	"",
+	"HORN_LOWRDER2",
+	"",
+	"HORN_XM15_1",
+	"",
+	"HORN_XM15_2",
+	"",
+	"HORN_XM15_3",
+	"",
+	"CMOD_AIRHORN_01",
+	"",
+	"CMOD_AIRHORN_02",
+	"",
+	"CMOD_AIRHORN_03",
+	""
+};
+
+inline const std::vector<std::string> vValues_ModSlotNames
+{
+	"CMOD_MOD_SPO",
+	"CMOD_BUM_1",
+	"CMOD_BUM_4",
+	"CMOD_MOD_SKI",
+	"CMOD_MOD_MUF",
+	"CMOD_MOD_CHA",
+	"CMOD_MOD_GRL",
+	"CMOD_MOD_HOD",
+	"CMOD_MOD_FEN",
+	"CMOD_MOD_FEN",
+	"CMOD_MOD_ROF",
+	"CMOD_MOD_ENG",
+	"CMOD_MOD_BRA",
+	"CMOD_MOD_TRN",
+	"CMOD_MOD_HRN",
+	"CMOD_MOD_SUS",
+	"CMOD_MOD_ARM",
+	"Unknown 17",
+	"CMOD_MOD_TUR",
+	"Unknown 19",
+	"Tyre Smoke",
+	"Unknown 21",
+	"CMOD_MOD_LGT",
+	"CMOD_MOD_WHEM",
+	"CMOD_WHE0_1",
+	"CMM_MOD_S0",//"Plateholder",
+	"CMM_MOD_S1",//"Vanity Plate",
+	"CMM_MOD_S2",//"Trim Design",
+	"CMM_MOD_S3",//"Ornament",
+	"CMM_MOD_S4",//"Dash",
+	"CMM_MOD_S5",//"Dial Design",
+	"CMM_MOD_S6",//"Doorcards",
+	"CMM_MOD_S7",//"Leather Seats",
+	"CMM_MOD_S8",//"Steering Wheel",
+	"CMM_MOD_S9",//"Column Shifter Lever",
+	"CMM_MOD_S10",//"Plaque",
+	"CMM_MOD_S11",//"Speakers",
+	"CMM_MOD_S12",//"Trunk",
+	"CMM_MOD_S13",//"Hydraulics",
+	"CMM_MOD_S14",//"Engine Block",
+	"CMM_MOD_S15",//"Air Filter",
+	"CMM_MOD_S16",//"Struts",
+	"CMM_MOD_S17",//"Arch Cover",
+	"CMM_MOD_S18",//"Aerials",
+	"CMM_MOD_S19",//"Trim",
+	"CMM_MOD_S20",//"Tank",
+	"CMM_MOD_S21",//"Windows",
+	"CMM_MOD_S22",//"Unknown 47",
+	"CMM_MOD_S23",//"Livery",
+};
